backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sse-starlette==2.0.0

# LLM providers
openai==1.12.0
llama-cpp-python==0.2.55

# Vector DB and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
numpy<2.0  # Required for chromadb compatibility

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# For markdown parsing
markdown==3.5.2

# Training dependencies (install with: pip install -r requirements-training.txt)
# These are optional - only needed for local model fine-tuning